Frontend: Simplify - compose with last pending
Instead of tracking a special change set called `current_change`, simply
compose every keystroke change set with the last element of the pending
log. Also, compose the inverted change set with the top of the undo
stack.

This is simpler and faster. We no longer need to keep a clone of the
prior `DocumentValue` before `current_change`.

Add a few other optimizations to avoid allocations.